Added setup.py installation script and helper Makefile.
The setup.py script uses the standard Python distutils package to build
and install our package. The script is designed to fail gracefully if a
dependency required for installation (not use) is missing.

A helper Makefile that wraps a few of the common invocations of the
setup script has also been provided. In particular, it provides a clean
target that cleans the temporary build files and any compiled modules.