Added Eclipse project files to .gitignore.
IDE-specific files and folders do not belong under version control.
Users of other IDEs should feel free to add the appropriate file
patterns to the top-level .gitignore file so that git ignores these
files. This commit provides file patterns for Eclipse, as if you have
created a project in the top-level directory of this repository.