Added documentation files as generated by sphinx-quickstart.
The documentation files were generated using Sphinx version 1.1.2. Note
that we will be using the MathJax extension to provide math support in
the HTML output. This has the advantage that no additional installation
is required by the users; however, it does require an internet
connection to view the rendered math.