Added several sections to installation section of documentation.
This section describes the various ways of installing PyRate. I
considered putting it in with the introduction, but I thought it was
important enough to be worth its own section, if for no other reason
than to make it easier to find. Again, this should be revised and
expounded upon as the codebase grows.