Added several physical constants to new pyrate.constants module.
The pyrate.constants module provides access to a number of physical
constants relevant in chemistry applications. The constants are available
as standard Python objects, as well as statically-typed floating-point
variables for use in Cythonized code. (A bit of cleverness involving the
globals() dict was required to achieve this with module-level variables,
as this seems to be a neglected feature of Cython.) The recommended way
to import this module is ``import pyrate.constants as constants`` so as
to not pollute the importing module's namespace.