Added section to reference manual for pyrate.constants module.
The reference manual will make extensive use of the autodoc extension in
Sphinx to extract the docstrings from the various modules and place them
in the HTML/LaTeX documentation in a readable way. This allows us to
avoid writing documentation in two places. Information not relevant for
module docstrings can always be placed in the .rst files directly, of
course.

The section was placed in a new chapter titled "Basic functions in
PyRate", as it is too short to be worth an entire chapter.