Added LinearRotor class to new pyrate.statmech.rotation module.
The LinearRotor class represents a rigid rotor for a linear molecule, i.e.
a rotor with a single moment of inertia (or rotational constant). Methods
are available for computing thermodynamic and statistical mechanical
properties of the rotor using either classical or quantum approaches.
The classical results are much faster to evaluate and nearly always
accurate except at temperatures well below room temperature.

Note that we only use the unit conversion functions in the LinearRotor
__init__() method. All other functions have required units for their
parameters and return values that are generally multiples of SI units
commonly used in chemistry applications.