Added IdealGasTranslation class to pyrate.statmech.translation module.
This class represents the translational degrees of freedom of an ideal
gas in three-dimensional space. The only parameter is the mass of the
gas molecules, as the volume is replaced by RT/P according to the ideal
gas law. Translational energy levels, as determined from a 3D
particle-in-a-box model, are very small compared to RT except at
temperatures approaching absolute zero, so only the classical limit
results have been implemented. (Like the NonlinearRotor class, a
NotImplementedError will be raised if you attempt to use the quantum
solution.)