Added pyrate.statmech.system.StatMech class for overall statmech models.
The StatMech class is useful for working with overall statistical
mechanical models of a molecular system involving many degrees of
freedom. It provides a list of these modes, along with attributes for
storing the ground-state energy and spin multiplicity, and methods for
evaluating the total thermodynamic and statistical mechanical
properties.