Added statistical mechanics chapter to reference manual.
This contains separate sections for each type of molecular degree of
freedom -- translation, rotation, vibration, and torsion -- as well as a
section for the combined statistical mechanical model of a molecular
system. Each section on a degree of freedom contains a summary of the
mathematics of the Schrodinger equation and its solution, usually in
both the quantum and classical limits.