Added HeatCapacityModel class to new pyrate.thermo.model module.
The new module is for collecting various thermodynamic models of heat
capacity (and enthalpy, entropy, etc.) The HeatCapacityModel class
provides a base class for such models; it stores common information such
as the minimum and maximum valid temperature of the model and a string
describing the model (e.g. where it came from). Several methods for
computing thermodynamic properties are defined, but raise
NotImplementedError so that derived classes remember to override them.