Added Wilhoit class to pyrate.thermo.model module.
The Wilhoit class represents a heat capacity model based on the Wilhoit
polynomial. The advantages of the Wilhoit polynomial are (1) guaranteed
correct heat capacity values at zero and infinite temperature and (2)
a surprisingly good recreation of the shape of the heat capacity profile
between these limits. The Wilhoit polynomial requires fewer parameters
than other heat capacity models, but is not the quickest to evaluate,
and is not as widely supported.