Added NASA class to pyrate.thermo.model module.
The NASA extension type provides an efficient Cython implementation of
a heat capacity model based around the (seven or nine coefficient) NASA
polynomial. The NASA polynomial is very quick to evaluate and is very
commonly supported in other chemistry codes, usually as a pair of NASA
polynomials per species. Dividing large temperature ranges into regions
is commonly done to improve the accuracy of the NASA polynomial.