Added methods for fitting Wilhoit polynomial to data.
You can do the fit either at a given value of the scaled temperature B
or use an optimization algorithm to choose the value of B that allows for
best fit of the data. The limits of heat capacity at zero and infinite
temperature are specified indirectly by providing the linearity of the
molecule, along with the number of vibrational and hindered rotor modes.