Added KineticsModel class to new pyrate.kinetics.model module.
The new module is for collecting various kinetics models, both
pressure-independent and pressure-dependent. The KineticsModel class
provides a base class for such models; it stores common information such
as the minimum and maximum valid temperature and pressure of the model
and a string describing the model (e.g. where it came from). Derived
classes must implement the getRateCoefficient() and isPressureDependent()
methods.