Added PDepArrhenius class to pyrate.kinetics.model module.
This class provides a means of representing k(T,P) values as a set of
Arrhenius expressions at various pressures, which are interpolated
between on a logarithmic pressure scale. This is sufficient for some
systems, but in general pressure-dependent kinetics behavior is strongly
non-Arrhenius, so there can be significant error introduced by using
this model. Note that Chemkin calls this the "PLOG" format.