Added Chebyshev class to pyrate.kinetics.model module.
The Chebyshev polynomial expression provides the best fit to k(T,P)
values, with the ability to approximate very complex pressure-dependent
effects. The tradeoff is that this is strictly a polynomial fit, so
there is no physical insight to the fitted parameters. Furthermore, you
must be careful to avoid significant nonphysical effects due to
polynomial behavior in the fit.