Added helper methods for computing quantities from momentum and position.
The added methods compute the centroids, center of mass, kinetic energy,
and ring polymer energy for a given momentum and position. They are very
tightly Cythonized in anticipation of being called a large number of
times during an RPMD simulation, to ensure that they are not the
bottleneck during the job execution.