Added RPMD function for constraining to dividing surface.
This is done using the SHAKE and RATTLE algorithms, which themselves
involve Lagrange multipliers. The constraining becomes a flag passed to
evolve(), which defaults to 0 (false) so that the constraint is turned
off.

The RATTLE algorithm actually has two steps; the first constrains the
position and adjusts the momentum so that the value of the dividing
surface is zero, while the second readjusts the momentum so that the
time derivative of the dividing surface is zero. The latter process does
not require iteration.