Added implementation of RPMD ring polymer time evolution.
The evolve() subroutine is used to take a single (optionally
constrained) time step. The position, momentum, potential, force, and
reaction coordinate are updated by integrating the classical Hamiltonian
for the ring polymer system. It turns out that the ring polymer term in
the Hamiltonian is best handled in normal mode space (see
doi:10.1063/1.3489925 for more); this is implemented in the
evolve_free_ring_polymer() subroutine.

The Hamiltonian also requires the potential of the system. The
get_potential() function is set up to call back to Python to evaluate a
user-provided function for this purpose. This is done purposefully, to
enable the user to use a quantum chemistry package to get energies if a
mapped PES is not available.