Added first implementation of RPMD recrossing factor calculation.
The compute_transmission_coefficient() method now has the
structure of a complete recrossing factor calculation, at least in terms
of the RPMD simulations being run. In particular, the process of starting
a parent trajectory, allowing it to evolve with the dividing surface
constraint and Andersen thermostat, and running child trajectories after
an initial equilibration period is now fleshed out. Hopefully all that
remains is the actual computation of the recrossing factor itself. Note
that the relax() method has been merged into this method.

Note that, in my desire to avoid global variables, the function
prototypes have become rather unwieldy. Since f2py does not support
derived types, we are probably stuck with this. However, the end user
will see the cleaner Python interface, so it's not so bad.