Added methods for saving bead and centroid positions vs. time to disk.
This is very important for visualizing the trajectories to ensure that
they are physically reasonable. We use the VMD software for visualization
purposes. Functionality for saving either or both of the beads and/or
centroids are provided. For a recrossing factor computation, you can
save either or both of the parent trajectory and/or a sampling of the
child trajectories.

Saving this data is very slow, so only do it when debugging or making
animations for presentations.