Small tweaks to the RPMD dividing surface constraint algorithm.
The tolerances for convergence were a bit too small for systems with a
large number of beads, so I raised them a bit. A better warning message
for if the maximum number of iterations is exceeded is also provided.
Note that I've never seen the algorithm take more than two iterations
unless it is going to iterate forever. (Do we even need iterations?)