Added pyrate.py main executable script with command-line argument parser.
The most common workflow for PyRate will be to construct an input file
describing a job to execute and use the ``pyrate.py`` script to execute
that job. The script currently parses and validates the command-line
arguments. Optional arguments are available for controlling the verbosity
of log messages and the directory output is written to.