Added logging functionality to main pyrate.py executable script.
This functionality initializes the logging system to write simultaneously
to both the console (with stderr redirected to stdout) and to a file on
disk. The log file always has at least verbose detail. After
initialization, a log header is printed containing information about the
program. Once the job is finished, a log footer is written to signal that
the job terminated.