Added functions for computing transition state dividing surface quantities.
These functions evaluate the value, gradient, and Hessian of the
transition state dividing surface. Note that this surface can be composed
of multiple equivalent structures when reaction-path degeneracies exist;
these functions are designed to handle this more general case.

The functions are written in Fortran 90 in anticipation of use in
RPMD calculations, which will need to be as efficient as possible.