Added helper math functions to new pyrate.rpmd._main f2py module.
The random() and randomn() functions return random numbers from a uniform
and normal distribution, respectively. They are designed so that we can
easily change the pseudo-random algorithm if desired; for now we are
using the random_number() intrinsic function.

The rfft() and irfft() methods compute the forward and inverse real FFT
of an array of data. The transformed data is in half-complex format.
These functions use the fftw3 library, so that is now a dependency.

These functions will be needed for conducting RPMD simulations to compute
reaction rate coefficients and other properties.