Added better subroutine for seeding the random number generator.
I was not calling the intrinsic random_seed() subroutine correctly, so
the code was actually using the same pseudo-random number sequence each
time, making it completely deterministic. The new seeding is implemented
in a new random_init() subroutine to use the system clock, as in an
example in the gfortran documentation.