Added umbrella potential and bias term to get_potential() subroutine.
These terms are only included if mode is set to 1. The umbrella potential
comes with a new parameter, kforce, representing the force constant of
the potential. Several function parameter lists were updated so that we
can pass kforce to the get_potential() subroutine at all required
times.

These terms will be needed for computing the static factor, part of the
procedure of using RPMD to compute reaction rate coefficients.