Added Python implementation of static factor calculation.
This calculation involves two parts. In the first, we equilibrate a
trajectory in each window, using the equilibrated positions from the
neighboring window to initialize the equilibration of the next window.
(This begins at xi = 1 and proceeds in both directions.) Once this is
done, a number of trajectories are sampled in each window in order to
determine the mean and variance of xi, which we will use to compute the
static factor via umbrella integration. (This last bit hasn't yet been
implemented.)

Note that we are using an umbrella potential and a bias term in this
calculation; these are turned on by setting mode = 1 when calling into
the Fortran codes.

This calculation seems like it will take quite a bit longer than the
recrossing factor calculation, but is also easier to parallelize...